company-project2-microservices/project2-microservices/api-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
mysql-connector-python==8.0.33
